requirements.txt
allure_python_commons==2.9.45
JPype1==1.4.1
pytest==7.1.2
requests==2.28.1
schedule==1.1.0
xlrd==2.0.1
xlutils==2.0.0
yagmail==0.15.293
